% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_digits}
\alias{get_digits}
\title{Figures out the minumum number of digits (up to four) to report after the decimal place}
\usage{
get_digits(x, max_digits = 4)
}
\arguments{
\item{x}{Numeric vector.}

\item{max_digits}{Numeric scalar: maximum number of digits.}
}
\value{

}
\description{
Figures out the minumum number of digits (up to four) to report after the decimal place
}
\examples{

get_digits(rnorm(20))


}
