% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{get_freq_tbl}
\alias{get_freq_tbl}
\title{Frequency Tables}
\usage{
get_freq_tbl(
  x,
  grouped = TRUE,
  start,
  width = 5,
  rf = TRUE,
  cf = FALSE,
  cp = FALSE,
  pr = FALSE,
  x_name = "x",
  solutions = TRUE,
  return_type = c("gt", "kable", "tbl"),
  sort_desc = FALSE,
  font_size = 12,
  ...
)
}
\arguments{
\item{x}{Numeric vector.}

\item{grouped}{Logical. Should this be a grouped frequency table?}

\item{start}{Numeric scalar. For ungrouped frequency tables, lowest value
represented in the table (default = \code{min(x)}). For grouped frequency
tables, the lower bound of the lowest interval.}

\item{width}{Numeric scalar. Width of interval for grouped frequency tables.
(Ignored if grouped is FALSE.)}

\item{rf}{Logical. Should a column for relative frequency be included
(default = TRUE)? Note: values will be rounded to the value of
round_interim.}

\item{cf}{Logical. Should a column for cumulative frequency be included
(default = FALSE)?}

\item{cp}{Logical. Should a column for cumulative percentage be included
(default = FALSE)? Note: values will be rounded to the value of
round_final.}

\item{pr}{Logical. Should a column for percentile rank be included (default =
FALSE)? Note: values will be rounded to the value of round_final.}

\item{x_name}{Character scalar. Name of x column in stub head (default =
"x").}

\item{solutions}{If \code{TRUE} (the default), renders the table as expected. If
\code{FALSE}, will show all of the requested columns, but will not display any
of the calculated values.}

\item{return_type}{Determines what type of object to return: either "tbl"
(for a \verb{[tibble::tibble()]} object), "gt" (for a \verb{[(gt::gt()]} object, the
default), or "kable" (for a \verb{[(knitr::kable()]} object with additional
formatting from \verb{[kableExtra::kableExtra()]}).}

\item{sort_desc}{Boolean scalar. If TRUE, will sort the frequency table in
descending order (by x). If FALSE (the default), will sort in ascending
order.}

\item{font_size}{Numeric scalar. Font size of table in points (as rendered by
\verb{[gt::gt()]}).}

\item{...}{Additional arguments to override default behaviors (see
\code{\link[=handcalcs_defaults]{handcalcs_defaults()}}).}
}
\value{
Depends on \code{return_type}: Either a \verb{[tibble::tibble()]} object
(\code{return_type="tbl"}), \verb{[gt::gt()]} object (\code{return_type="gt"}), or a
\verb{[knitr::kable()]} object (\code{return_type="kable"}).
}
\description{
Produces both ungrouped and grouped frequency tables (as a \verb{[gt::gt()]} table
object) from a set of raw data.
}
\details{
By default, just calculates frequencies. Optionally, can include additional
columns for relative frequency (\code{rf}), cumulative frequency (\code{cf}),
cumulative percentage (\code{cp}), and percentile rank (\code{pr}). Additional
customization options for the \verb{[gt::gt()]} table object are also available.

Note that only numeric vectors can be used with the grouped option. Also note
that numeric vectors are sorted descending, whereas character vectors and
factors are sorted ascending.
}
\examples{

x <- sample(x = 1:20, size = 20, replace = TRUE)

# Ungrouped:
get_freq_tbl(x, grouped = FALSE, round_interim = 3)
get_freq_tbl(x, grouped = FALSE, cf = TRUE, cp = TRUE, pr = TRUE, round_final = 2)

# Grouped:
get_freq_tbl(x, grouped = TRUE)
get_freq_tbl(x, grouped = TRUE, width = 3, start = 0, cf = TRUE, cp = TRUE, pr = TRUE, round_final = 2)

# Also works on non-numeric data:
x <- sample(x = LETTERS, size = 20, replace = TRUE)
get_freq_tbl(x, grouped = FALSE)
get_freq_tbl(x, grouped = FALSE, cf = TRUE, cp = TRUE, pr = TRUE)

# When x is a factor, will include zeros for missing levels
x <- factor(sample(x = LETTERS, size = 20, replace = TRUE), levels = LETTERS)
get_freq_tbl(x, grouped = FALSE)
get_freq_tbl(x, grouped = FALSE, cf = TRUE, cp = TRUE, pr = TRUE)

}
