% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{get_handcalcs_opts}
\alias{get_handcalcs_opts}
\title{Generates list of options, overriding the global options if specified.}
\usage{
get_handcalcs_opts(...)
}
\arguments{
\item{...}{Default options to override}
}
\value{

}
\description{
get_handcalcs_opts() is a convenience function for setting options(handcalcs = .). (See
the examples for usage.) This is also used internally for parameters to be
passed to individual solution functions to temporarily override the global
default.
}
\examples{
# With no arguments, simply returns a complete list of the current global options
get_handcalcs_opts()

# When arguments are specified, modifies the list accordingly
get_handcalcs_opts(round_interim = 5)

# Most common usage will be to update the package defaults in global usage:
options(handcalcs = get_handcalcs_opts(round_interim = 5))

}
