% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handcalcs.R
\docType{package}
\name{handcalcs}
\alias{handcalcs}
\title{handcalcs: A package for mimicking hand calculations and generating solutions
when teaching statistics}
\description{
The handcalcs package covers numerous basic statistical functions, including
measures of central tendency, variability, z-tests, t-tests, ANOVA,
regression, and correlation.
}
