% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handcalcs.R
\name{handcalcs_defaults}
\alias{handcalcs_defaults}
\title{Default behaviors for handcalcs functions}
\usage{
handcalcs_defaults()
}
\value{
Named list of default values
}
\description{
Sets default behaviors for handcalcs functions related to rounding for
interim/final calculations and handling of LaTeX solution string. Any of the
defaults listed below can be changed by setting the corresponding value of
the named list ("handcalcs") in \code{options()}.
}
\details{
\describe{

\item{\code{round_interim}}{Number of digits past the decimal to round
interim calculations. (Default = 4.)}

\item{\code{round_final}}{Number of digits past the decimal to report final
result. (Default = 4.)}

\item{\code{add_math}}{Enclose solution string in LaTeX math block (\verb{$$})? If
set to \code{FALSE}, must manually include the math block tags in order for the
solution to render properly. (Default = TRUE.)}

\item{\code{add_aligned}}{Enclose solution in LaTeX aligned block
(\verb{\\begin\{aligned\}}). (Default = TRUE.)}

\item{\code{use_aligned}}{Use "&=" instead of "=" in solution string to
generate aligned equations. (Default = TRUE.)}

}
}
\examples{
handcalcs_defaults()

}
