% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es.R
\name{interpret_d}
\alias{interpret_d}
\title{Interpret Cohen's d}
\usage{
interpret_d(d, ...)
}
\arguments{
\item{d}{Numeric. Value of Cohen's d (effect size in standard deviation
units).}

\item{...}{Additional arguments to override default behaviors (see
\code{\link[=handcalcs_defaults]{handcalcs_defaults()}}}
}
\value{
Interpretation as a character string: small (d <= 0.2), medium (0.2 >
d <= 0.8), or large (d > 0.8)
}
\description{
Interpret Cohen's d
}
\examples{
interpret_d(seq(0, 1, by = 0.1))
}
