% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{show_raw_data}
\alias{show_raw_data}
\title{Show Raw Data}
\usage{
show_raw_data(
  x,
  width = 10,
  sort_x = c("none", "asc", "desc"),
  font_size = 12,
  ...
)
}
\arguments{
\item{x}{Numeric vector.}

\item{width}{Numeric scalar. Width of table to be produced.}

\item{sort_x}{Character. Should data be sorted before being displayed? Can
specify ascending or descending sort.}

\item{font_size}{Numeric scalar. Font size (in pixels).}

\item{...}{Additional arguments to override default behaviors (see
\code{\link[=handcalcs_defaults]{handcalcs_defaults()}}).}
}
\value{
A gt object.
}
\description{
Produces a simple table (\verb{[gt::gt()]} object) of raw data by putting the data
into rows (using specified \code{width}). Note: raw data will be rounded to
the value of round_interim (either supplied as an option, or consulting the
global default value).
}
\examples{
# Generate sample data:
x = sample(1:20, 20, replace = TRUE)
show_raw_data(x, sort = 'asc')

# Note rounding behavior:
x = rnorm(20)
show_raw_data(x, round_interim = 2)

# If width is not a multiple of the length of x, will pad with empty cells:
show_raw_data(x, width = 7)

}
