% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{solve_auc_to_z}
\alias{solve_auc_to_z}
\title{Calculating Values of Z from Area Under the Normal Curve}
\usage{
solve_auc_to_z(a, type = c("above", "below", "between", "tails"), ...)
}
\arguments{
\item{a}{Numeric scalar. Value of area specified by \code{type}. Must be between 0
and 1.}

\item{type}{Character. Determines what type of area is being provided:
\code{above}, \code{below}, \code{between}, or \code{tails}.}

\item{...}{Additional arguments to override default behaviors (see
\code{\link[=handcalcs_defaults]{handcalcs_defaults()}}).}
}
\value{
Returns a list with the provided value (\code{a}), any interim areas
calculated (\code{a_above}, \code{a_below}, \code{a_between}, \code{a_1tail}, \code{a_2tails},
depending on the type of area provided), the final value (\code{z} for \code{type = above} and \code{type = below}; \code{z1} and \code{z2} for \code{type = between} and \code{type = tails}), and the solution string (\code{solution}).
}
\description{
Takes a given area under the curve (AUC) and calculates the corresponding
value of \code{z} (e.g., a \emph{z}-score). The type argument determines what area is
being provided:
}
\details{
\describe{

\item{\code{above}}{Value of \code{a} is the area above the value of \code{z} to be
calculated.}

\item{\code{above}}{Value of \code{a} is the area below the value of \code{z} to be
calculated.}

\item{\code{between}}{Value of \code{a} is the area between two values of \code{z}.
Always returns a positive and negative value of \code{z}.}

\item{\code{tails}}{Value of \code{a} is the area more extreme than two values of
\code{z} (i.e., in the tails). (Note: this is the total area in the tails, not
just in one tail.) Always returns a positive and negative value of \code{z}.}

}

Note that \emph{z}-scores are rounded to the value of \code{round_z} instead of the
value of \code{round_interim} or \code{round_final} (see \code{\link[=handcalcs_defaults]{handcalcs_defaults()}}).
}
\examples{

# Calculates the area above and below a single z value.
solve_auc_to_z(a = 0.25, 'above')
solve_auc_to_z(a = 0.25, 'below')

solve_auc_to_z(a = 0.68, 'between')
solve_auc_to_z(a = 0.32, 'tails')


}
