% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{solve_ci_t_indep}
\alias{solve_ci_t_indep}
\alias{ci_t_indep_formula}
\title{Confidence Intervals of Difference Between Independent Means (using \emph{t}
distribution)}
\usage{
solve_ci_t_indep(
  M_1,
  M_2,
  s_1,
  s_2,
  SD_1,
  SD_2,
  s_p2,
  n_1,
  n_2,
  df,
  s_M_diff,
  level = 0.95,
  ...
)

ci_t_indep_formula(level = 0.95, ...)
}
\arguments{
\item{M_1, M_2}{Numeric scalars. Means of samples 1 and 2.}

\item{s_1, s_2, SD_1, SD_2}{Numeric scalars. Standard deviations of samples 1
and 2. May be named either \code{s_1} and \code{s_2} or \code{SD_1} and \code{SD_2}. Required
if not providing pooled variance (\code{s_p2}) or the standard error of the
differences (\code{s_M_diff}).}

\item{s_p2}{Numeric scalar. Pooled variance. Will be calculated from sample
standard deviations (\code{s_1}, \code{s_2}) and sample sizes (\code{n_1} and \code{n_2}) if
those values are provided.}

\item{n_1, n_2}{Numeric scalars. Sample sizes of samples 1 and 2. Required if
s_M_diff is not provided. Also required if \code{df} is not provided.}

\item{df}{Numeric scalar. Degrees of freedom (df, \eqn{df = n_1 + n_2 - 1}).
If not provided, will be calculated from \code{n_1} and \code{n_2}. calculated.}

\item{s_M_diff}{Numeric scalar. Standard error of the difference of the
means. If not provided, will be calculated (using \code{\link[=solve_s_M_diff]{solve_s_M_diff()}}) from
sample sizes (\code{n_1} and \code{n_2}) and standard deviations (\code{s_1}, \code{s_2}) and
included in the solution string.}

\item{level}{Numeric scalar. Confidence level; defaults to 0.95 (95\%
confidence intervals).}

\item{...}{Additional arguments to override default behaviors (see
\code{\link[=handcalcs_defaults]{handcalcs_defaults()}}}
}
\value{
\code{solve_ci_t_indep()} returns a list with all provided values (\code{M_1},
\code{M_2}, \code{s_1}, \code{s_2}, \code{SD_1}, \code{SD_2}, \code{n_1}, \code{n_2}, \code{df}, \code{s_M_diff},
\code{level}), the interim calculations (\code{s_M_diff}, \code{t_crit}, \code{marg_err}), the
final value (\code{CI_lower}, \code{CI_upper}, as well as a named vector containing
both values \code{CI}), the solution string (\code{solution}), and the bare formula
(\code{formula}) in LaTeX format. \code{ci_t_indep_formula()} returns just the bare
formula in LaTeX format as a character string.
}
\description{
Calculates confidence intervals for the difference between two independent
sample means: \eqn{(M_{1} - M_{2}) \pm (t_{1 - \alpha/2})(s_{M_{1} -
M_{2}})}. May either provide \code{s_M_diff}, or that value can be calculated by
providing the sample sizes (\code{n_1} and \code{n_2}) and either the pooled variance
(\code{s_p2}) or the two sample standard deviations (\code{s_1} and \code{s_2}).
}
\details{
You must provide either \code{df} or the two sample sizes (\code{n_1} and \code{n_2}). If
sample sizes are provided, \code{df} will be calculated (\eqn{df = n_1 + n_2 -
1}). If both are provided, this allows for the possibility of errors that
produce incorrect results. However, it also allows for the possibility of
using a different value of \code{df} than \code{n_1 + n_2 - 1}, for instance of
requiring use of a statistical table that does not have that exact value of
\code{df} listed.

Note that the critical value of \emph{t} (\code{t_crit}) is rounded to the value of
\code{round_t} instead of the value of \code{round_interim} or \code{round_final} (see
\code{\link[=handcalcs_defaults]{handcalcs_defaults()}}).
}
\examples{

<<<<<<< HEAD
solve_ci_t_indep(M_1 = 55, M_2 = 50, s_M_diff = 1, df = 98)

# Defaults to 95\% CIs, but other levels can be specified:
solve_ci_t_indep(M_1 = 55, M_2 = 50, s_M_diff = 1, df = 98, level = 0.99)

# If s_M_diff is not provided, can be calculated from sample sizes and either
# pooled variance (s_p2):
solve_ci_t_indep(M_1 = 50, M_2 = 45, n_1 = 50, n_2 = 50, s_p2 = 0.5)

# or standard deviations (s_1 and s_2)
solve_ci_t_indep(M_1 = 50, M_2 = 45, n_1 = 50, n_2 = 50, s_1 = 1.5, s_2 = 2.0)
=======
solve_ci_t_indep(M = 55, s_M_diff = 1, df = 99)

# Defaults to 95\% CIs, but other levels can be specified:
solve_ci_t_indep(M = 55, s_M_diff = 1, df = 99, level = 0.99)

# Will calculate s_M_diff if s and n (or df) are provided, and include the
# calculation in the solution string
solve_ci_t_indep(M = 55, s = 10, n = 100)

# Note: if s_M_diff is provided, it takes precedence; s will be recalculated
solve_ci_t_indep(M = 55, s = 10, n = 100, s_M_diff = 4)

# If mu is provided, will return the CI of the mean difference (M - mu).
solve_ci_t_indep(M = 55, s_M_diff = 1, n = 100, mu = 50)
>>>>>>> be61bc195412aca1a4d0617fb50c07563a278533

# If you just want the formula:
ci_t_indep_formula()

<<<<<<< HEAD
=======
# To get the formula for the mean difference, supply a non-null value for mu:
ci_t_indep_formula(mu = TRUE)
>>>>>>> be61bc195412aca1a4d0617fb50c07563a278533

}
