% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{solve_ci_t_paired}
\alias{solve_ci_t_paired}
\alias{ci_t_paired_formula}
\title{Confidence Intervals Around Mean Difference Score (using \emph{t} distribution)}
\usage{
solve_ci_t_paired(
  M_D,
  s_D,
  SD_D,
  n,
  df,
  s_M_D,
  mu_D = NULL,
  paired = FALSE,
  level = 0.95,
  ...
)

ci_t_paired_formula(mu_D = NULL, level = 0.95, ...)
}
\arguments{
\item{M_D}{Numeric scalar. Mean of the difference scores.}

\item{s_D, SD_D}{Numeric scalar. Standard deviation of the difference scores.
May be named either \code{s_D} or \code{SD_D}. Required if s_M_D is not provided.}

\item{n, df}{Numeric scalar. Must supply either sample size (n) or degrees of
freedom (df, \eqn{df = n - 1}). Whichever argument is not provided will be
calculated.}

\item{s_M_D}{Numeric scalar. Standard error of the mean of the difference
scores. If not provided, will be calculated from \code{s_D} (or \code{SD_D}) and \code{n}
(using \code{\link[=solve_s_M_D]{solve_s_M_D()}}) and included in the solution string.}

\item{mu_D}{Numeric scalar. Predicted mean difference score under the null
hypothesis. If provided, will return the confidence interval around the
mean difference (\eqn{M_{D} - \mu_{D}}).}

\item{level}{Numeric scalar. Confidence level; defaults to 0.95 (95\%
confidence intervals).}

\item{...}{Additional arguments to override default behaviors (see
\code{\link[=handcalcs_defaults]{handcalcs_defaults()}}}
}
\value{
\code{solve_ci_t_paired()} returns a list with the provided values
(\code{M_D},\code{s_D}, \code{SD_D}, \code{n}, \code{s_M_D}), the interim calculations (\code{s_M_D},
\code{t_crit}, \code{marg_err}), the final value (\code{CI_lower}, \code{CI_upper}, as well as
a named vector containing both values \code{CI}), the solution string
(\code{solution}), and the bare formula (\code{formula}) in LaTeX format.
\code{ci_t_paired_formula()} returns just the bare formula in LaTeX format as a
character string.
}
\description{
Calculates confidence intervals around the mean difference score \eqn{M_{D}
\pm (t_{1 - \alpha/2})(s_{M_{D}})}. May either provide \code{s_M_D}, or that value
can be calculated from \code{SD_D} and \code{n}.
}
\details{
Optionally, provide the value of \code{mu_D} (\eqn{mu_{D}}), the value of the mean
difference predicted by the null hypothesis), which will return the
confidence interval around the mean difference: \eqn{(M_{D} - \mu_{D}) \pm
(t_{1 - \alpha/2})(s_{M_{D}})}.

You must provide either \code{df} or \code{n}. If one is provided, the other value will
be calculated (\eqn{df = n - 1}). If both are provided, this allows for the
possibility of errors that produce incorrect results. However, it also allows
for the possibility of using a different value of \code{df} than \code{n - 1}, for
instance of requiring use of a statistical table that does not have that
exact value of \code{df} listed.

Note that the critical value of \emph{t} (\code{t_crit}) is rounded to the value of
\code{round_t} instead of the value of \code{round_interim} or \code{round_final} (see
\code{\link[=handcalcs_defaults]{handcalcs_defaults()}}).
}
\examples{

solve_ci_t_paired(M_D = 5, s_M_D = 1, df = 99)

# Defaults to 95\% CIs, but other levels can be specified:
solve_ci_t_paired(M_D = 5, s_M_D = 1, df = 99, level = 0.99)

# Will calculate s_M_D if s_D and n (or df) are provided, and include the
# calculation in the solution string
solve_ci_t_paired(M_D = 5, s_D = 10, n = 100)

# Note: if s_M_D is provided, it takes precedence; s_D will be recalculated
solve_ci_t_paired(M_D = 5, s_D = 10, n = 100, s_M_D = 4)

# If mu is provided, will return the CI of the mean difference (M_D - mu_D).
solve_ci_t_paired(M_D = 5, s_M_D = 1, n = 100, mu_D = 0)

# If you just want the formula:
ci_t_paired_formula()

# To get the formula for the mean difference, supply a non-null value for mu:
ci_t_paired_formula(mu_D = TRUE)

}
