% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es.R
\name{solve_cohens_d_indep_t}
\alias{solve_cohens_d_indep_t}
\alias{cohens_d_indep_t_formula}
\title{Effect site (Cohen's d) for independent-samples \emph{t}-test}
\usage{
solve_cohens_d_indep_t(M_1, M_2, SD2_p, SD_p, interpret = FALSE, ...)

cohens_d_indep_t_formula(...)
}
\arguments{
\item{M_1}{Numeric scalar. Mean of sample 1.}

\item{M_2}{Numeric scalar. Mean of sample 2.}

\item{SD2_p}{Numeric scalar. Pooled variance. Required if SD_p is not
provided.}

\item{SD_p}{Numeric scalar. Pooled standard deviation. Required if SD2_p is
not provided.}

\item{interpret}{Logical scalar. Should the interpretation (as produced by
\code{\link[=interpret_d]{interpret_d()}} be included in the solution string?)}

\item{...}{Additional arguments to override default behaviors (see
\code{\link[=handcalcs_defaults]{handcalcs_defaults()}}}
}
\value{
\code{solve_cohens_d_indep_t()} returns a list with the provided values
(\code{M_1}, \code{M_2}, \code{SD2_p}, \code{SD_p}), the interim calculations (\code{M_diff}), the
final value (\code{d}), the solution string (\code{solution}), and the bare formula
(\code{formula}) in LaTeX format. \code{cohens_d_indep_t_formula()} returns just the
bare formula in LaTeX format as a character string.
}
\description{
Calculates and presents solution string for effect site (Cohen's d) comparing
sample means from independent samples: the difference between sample means
(\code{M_1} and \code{M_2}) in units of the pooled standard deviation (\code{SD_p}), which
can also be given as the pooled variance (\code{SD2_p}). One of those two
arguments must be provided, but not both.
}
\examples{

# Must provide either the pooled SD or the pooled variance:
solve_cohens_d_indep_t(M_1 = 6, M_2 = 5, SD_p = 2)
solve_cohens_d_indep_t(M_1 = 6, M_2 = 5, SD2_p = 4)

# Include the interpretation of the effect size:
solve_cohens_d_indep_t(M_1 = 6, M_2 = 5, SD_p = 2, interpret = TRUE)

# You must provide either the pooled variance (SD2_p) or the pooled
# standard deviation (SD_p), but not both.
\dontrun{solve_cohens_d_indep_t(M_1 = 6, M_5 = 2)}
\dontrun{solve_cohens_d_indep_t(M_1 = 6, M_5 = 2, SD2_p = 4, SD_p = 2)}

# If you just want the formula:
cohens_d_indep_t_formula()

}
