% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es.R
\name{solve_cohens_d_one_t}
\alias{solve_cohens_d_one_t}
\alias{cohens_d_one_t_formula}
\title{Effect site (Cohen's d) for one-sample \emph{t}-test}
\usage{
solve_cohens_d_one_t(M, mu, SD, interpret = FALSE, ...)

cohens_d_one_t_formula(...)
}
\arguments{
\item{M}{Numeric scalar. Sample mean.}

\item{mu}{Numeric scalar. Population mean.}

\item{SD}{Numeric scalar. Sample standard deviation.}

\item{interpret}{Logical scalar. Should the interpretation (as produced by
\code{\link[=interpret_d]{interpret_d()}} be included in the solution string?)}

\item{...}{Additional arguments to override default behaviors (see
\code{\link[=handcalcs_defaults]{handcalcs_defaults()}}}
}
\value{
\code{solve_cohens_d_one_t()} returns a list with the provided values
(\code{M}, \code{mu}, \code{SD}), the interim calculations (\code{M_diff}), the final
value (\code{d}), the solution string (\code{solution}), and the bare formula
(\code{formula}) in LaTeX format. \code{cohens_d_one_t_formula()} returns just the
bare formula in LaTeX format as a character string.
}
\description{
Calculates and presents solution string for effect site (Cohen's d) comparing
a single sample to a population mean when the population standard deviation
is \emph{not} known (and must be estimated): difference between observed sample mean (\code{M}) and population mean
(\code{mu}) in units of the sample standard deviation (\code{SD}).
}
\examples{

solve_cohens_d_one_t(M = 6, mu = 5, SD = 2)

# Include the interpretation of the effect size:
solve_cohens_d_one_t(M = 6, mu = 5, SD = 2, interpret = TRUE)

# If you just want the formula:
cohens_d_one_t_formula()
}
