% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es.R
\name{solve_cohens_d_one_z}
\alias{solve_cohens_d_one_z}
\alias{cohens_d_one_z_formula}
\title{Effect size (Cohen's d) for one-sample \emph{z}-test}
\usage{
solve_cohens_d_one_z(M, mu, sigma, interpret = FALSE, ...)

cohens_d_one_z_formula(...)
}
\arguments{
\item{M}{Numeric scalar. Sample mean.}

\item{mu}{Numeric scalar. Population mean.}

\item{sigma}{Numeric scalar. Population standard deviation.}

\item{interpret}{Logical scalar. Should the interpretation (as produced by
\code{\link[=interpret_d]{interpret_d()}} be included in the solution string?)}

\item{...}{Additional arguments to override default behaviors (see
\code{\link[=handcalcs_defaults]{handcalcs_defaults()}}}
}
\value{
\code{solve_cohens_d_one_z()} returns a list with the provided values
(\code{M}, \code{mu}, \code{sigma}), the interim calculations (\code{M_diff}), the final
value (\code{d}), the solution string (\code{solution}), and the bare formula
(\code{formula}) in LaTeX format. \code{cohens_d_one_z_formula()} returns just the
bare formula in LaTeX format as a character string.
}
\description{
Calculates and presents solution string for effect size (Cohen's d) comparing
a single sample to a population mean when the population standard deviation
is known: difference between observed sample mean (\code{M}) and population mean
(\code{mu}) in units of the population standard deviation (\code{sigma}).
}
\examples{

solve_cohens_d_one_z(M = 6, mu = 5, sigma = 2)

# Include the interpretation of the effect size:
solve_cohens_d_one_z(M = 6, mu = 5, sigma = 2, interpret = TRUE)

# If you just want the formula:
cohens_d_one_z_formula()
}
