% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_scores.R
\name{solve_diff_score}
\alias{solve_diff_score}
\alias{diff_score_formula}
\title{Difference Scores}
\usage{
solve_diff_score(x_1, x_2, sym_x = "X", abbrev_at = 5, ...)

diff_score_formula(sym_x = "X", ...)
}
\arguments{
\item{x_1}{Numeric vector of values.}

\item{x_2}{Numeric vector of values.}

\item{sym_x}{Character scalar. Symbol to represent x in formula (default:
"X").}

\item{abbrev_at}{Numeric scalar. At what length of \code{x_1} should it abbreviate
the list of values within the solution using an ellipsis? If the length of
\code{x_1} is greater than or equal to \code{abbrev_at}, it will abbreviate the
sequence (e.g., "1, 2, ..., 6"). Otherwise, will show the entire sequence
(e.g., "1, 2, 3, 4, 5, 6).}

\item{...}{Additional arguments to override default behaviors (see
\code{\link[=handcalcs_defaults]{handcalcs_defaults()}}).}
}
\value{
\code{solve_diff_score()} returns a list with the provided values (\code{x_1},
\code{x_2}), the final value (the vector of difference scores, \code{x_D}), the
solution string (\code{solution}), and the bare formula (\code{formula}) in LaTeX
format. \code{diff_score_formula()} returns just the bare formula in LaTeX
format as a character string.
}
\description{
Calculates and provides solutions for calculating difference scores
(\eqn{X_{D} = X_{2} - X_{1}}).
}
\examples{

x_1 <- rnorm(10, mean = 0)
x_2 <- rnorm(10, mean = 1)
solve_diff_score(x_1 = x_1, x_2 = x_2)

# Can use a different symbol for X:
solve_diff_score(x_1 = x_1, x_2 = x_2, sym_x = "Y")

# If you just want the bare formula as a string, use the formula function():
diff_score_formula()

}
