% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc.R
\name{solve_LSD}
\alias{solve_LSD}
\alias{LSD_formula}
\title{Fisher's LSD (Least Significant Difference) Test}
\usage{
solve_LSD(M_1, M_2, n_1, n_2, MS_error, ...)

LSD_formula(...)
}
\arguments{
\item{M_1}{Numeric scalar. Mean of sample 1.}

\item{M_2}{Numeric scalar. Mean of sample 2.}

\item{n_1}{Numeric scalar. Sample size of sample 1.}

\item{n_2}{Numeric scalar. Sample size of sample 2.}

\item{MS_error}{Numeric scalar. Mean Squared Error term from the ANOVA.}

\item{...}{Additional arguments to override default behaviors (see
\code{\link[=handcalcs_defaults]{handcalcs_defaults()}}}
}
\value{
\code{solve_t_obt()} returns a list with the provided values (\code{M_1},
\code{M_2}, \code{n_1}, \code{n_2}, \code{MS_error}), the interim calculations (\code{M_diff},
\code{MS_error_n1}, \code{MS_error_n2}, \code{Sum_MS_error_n}, \code{s_M_diff2}, \code{s_M_diff}),
the final value (\code{t_obt}), the solution string (\code{solution}), and the bare
formula (\code{formula}) in LaTeX format (note: empty values are removed from
the list). \code{Q_obt_formula()} returns just the bare formula in LaTeX format
as a character string.
}
\description{
Calculates and provides solutions for Fisher's LSD, intended as a means to
conduct planned comparisons after a significant ANOVA. It is essentially the
same as an independent-samples t-test, except that it uses \eqn{MS_{error}}
in place of the pooled variance \eqn{s_{p}^2}.
}
\details{
Requires both group means (\code{M_1}, \code{M_2}) and sample sizes (\code{n_1}, \code{n_2}), as
well as the error term from the ANOVA (\code{MS_error}).

The result (\code{t_obt}) is calculated as
\eqn{\frac{|M_1-M_2|}{\sqrt{(MS_{error}/n_1 + MS_{error}/n_2)}}}. To
determine significance, either use the built-in R function \code{\link[=pt]{pt()}} or compare
to the critical value using the \code{\link[=solve_t_crit]{solve_t_crit()}} function.
}
\examples{

solve_LSD(M_1 = 5, M_2 = 4, n_1 = 52, n_2 = 50, MS_error = 2.5)

}
