% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centraltendency.R
\name{solve_mean}
\alias{solve_mean}
\title{Arithmetic mean: calculate value (with rounding) and present solutions}
\usage{
solve_mean(x, sub = "", sym = ifelse(nchar(sub) == 1, sub, "X"), ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{sub}{Character scalar. Name of numeric vector (reported as subscript in
solutions). Leave empty to report no subscript.}

\item{sym}{Character scalar. Symbol to represent x in formula (default: "X").
Only one character allowed.}

\item{...}{Additional arguments to override default behaviors (see
\code{\link{handcalcs_defaults}}).}
}
\value{
A list with the interim calculations (\code{SumX}, \code{n}), the
final value (\code{M}), and the solution string (\code{solution}) in LaTeX format.
}
\description{
Arithmetic mean: calculate value (with rounding) and present solutions
}
\examples{
x <- sample(x = 1:10, size = 20, replace = TRUE)
solve_mean(x)

# Note that the value calculated by solve_mean(x) will be different than the
# value calculated by mean(x) when rounding occurs.
(x <- rnorm(20))
solve_mean(x, sub = "Y")

# The value of sub can be long, but not sym.
solve_mean(x, sub = "Donuts")

}
