% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centraltendency.R
\name{solve_median}
\alias{solve_median}
\title{Median: calculate value (with rounding) and present solutions}
\usage{
solve_median(x, sub = "", ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{sub}{Character scalar. Name of numeric vector (reported as subscript in
solutions). Leave empty to report no subscript.}

\item{...}{Additional arguments to override default behaviors (see
\code{\link{handcalcs_defaults}}).}
}
\value{
A list with the median value (\code{Med}) and the solution
(\code{solution}) in LaTeX format.
}
\description{
Median: calculate value (with rounding) and present solutions
}
\examples{
#' x <- sample(x = 1:10, size = 20, replace = TRUE)
solve_median(x, sub = "Y")

# Note that the raw data values get rounded by solve_median(x), so the exact
# value of the median will differ from the value calculated by median(x)
# when there are a lot of digits past zero.
(x <- rnorm(20))
Med <- solve_median(x, round_interim = 2)$Med
Med == median(x)

}
