% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical.R
\name{solve_Q_crit}
\alias{solve_Q_crit}
\title{Critical values of \emph{Q} (Studentized Range Statistic)}
\usage{
solve_Q_crit(alpha = 0.05, df_error, k, ...)
}
\arguments{
\item{alpha}{Numeric scalar. Value of alpha (probability of a Type I error).
Defaults to 0.05.}

\item{df_error}{Numeric scalar. Degrees of freedom.}

\item{k}{Numeric scalar. Number of groups in design.}

\item{...}{Additional arguments to override default behaviors (see
\code{\link[=handcalcs_defaults]{handcalcs_defaults()}}).}
}
\value{
Returns a list with the given values (\code{alpha}, \code{df}, \code{k}), the final
value (\code{Q_crit}), and the solution string (\code{solution}).
}
\description{
Calculates the critical value of \emph{Q} (the Studentized Range Statistic for
multiple comparisons) for a given \eqn{\alpha}, degrees of freedom for error
(\code{df_error}) from the ANOVA, and number of groups being compared (\code{k}), along
with the solution string in LaTeX format.
}
\details{
This is used when conducting \emph{post hoc} tests after a significant ANOVA,
considering each mean difference, taking into consideration the total number
of comparisons to maintain an overall Type I error rate (\code{alpha}). The
critical value would be compared against the value of Q_obt from the
\verb{[solve_Q_obt()]} function for each comparison.
}
\examples{

solve_Q_crit(alpha = 0.01, df_error = 100, k = 3)
solve_Q_crit(alpha = 0.01, df_error = 100, k = 4)
solve_Q_crit(alpha = 0.01, df_error = 100, k = 5)

# In some cases, you may wish to change the default value for round_final:
solve_Q_crit(alpha = 0.05, df = 100, k = 3, round_final = 4)

}
