% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posthoc.R
\name{solve_Q_obt}
\alias{solve_Q_obt}
\alias{Q_obt_formula}
\title{Tukey's HSD Post Hoc Test (Studentized Range Statistic)}
\usage{
solve_Q_obt(M_1, M_2, n_1, n_2, MS_error, ...)

Q_obt_formula(...)
}
\arguments{
\item{M_1}{Numeric scalar. Mean of sample 1.}

\item{M_2}{Numeric scalar. Mean of sample 2.}

\item{n_1}{Numeric scalar. Sample size of sample 1.}

\item{n_2}{Numeric scalar. Sample size of sample 2.}

\item{MS_error}{Numeric scalar. Mean Squared Error term from the ANOVA.}

\item{...}{Additional arguments to override default behaviors (see
\code{\link[=handcalcs_defaults]{handcalcs_defaults()}}}
}
\value{
\code{solve_q_obt()} returns a list with the provided values (\code{M_1},
\code{M_2}, \code{n_1}, \code{n_2}, \code{MS_error}), the interim calculations (\code{M_diff},
\code{MS_error_n1}, \code{MS_error_n2}, \code{Sum_MS_error_n}, \code{s_M_diff2}, \code{s_M_diff}),
the final value (\code{Q_obt}), the solution string (\code{solution}), and the bare
formula (\code{formula}) in LaTeX format (note: empty values are removed from
the list). \code{Q_obt_formula()} returns just the bare formula in LaTeX format
as a character string.
}
\description{
Calculates and provides solutions for Tukey's HSD using the Studentized Range
Statistic (\code{Q_obt}). This is primarily useful in calculating post hoc tests
making all possible comparisons among \code{k} groups after a significant ANOVA,
while maintaining the overall Type I error rate (\code{alpha}). Requires both
group means (\code{M_1}, \code{M_2}) and sample sizes (\code{n_1}, \code{n_2}), as well as the
error term from the ANOVA (\code{MS_error}).
}
\details{
The result (\code{Q_obt}) is calculated as
\eqn{\frac{|M_1-M_2|}{\sqrt{(MS_{error}/n_1 + MS_{error}/n_2)(1/2)}}}. To
determine significance, either use the built-in R function \code{\link[=ptukey]{ptukey()}} or
compare to the critical value using the \code{\link[=solve_Q_crit]{solve_Q_crit()}} function (which
itself makes use of the built-in \code{\link[=qtukey]{qtukey()}} function).
}
\examples{

solve_Q_obt(M_1 = 5, M_2 = 4, n_1 = 52, n_2 = 50, MS_error = 2.5)

}
