% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.R
\name{solve_range}
\alias{solve_range}
\alias{range_formula}
\title{Range: calculate value (with rounding) and present solutions}
\usage{
solve_range(x, sub = "", sym = "X", ...)

range_formula(sub = "", sym = "X", ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{sub}{Character scalar. Name of numeric vector (reported as subscript in
solutions). Leave empty to report no subscript.}

\item{sym}{Character scalar. Symbol to represent x in formula (default: "X").
Only one character allowed.}

\item{...}{Additional arguments to override default behaviors (see
\code{\link{handcalcs_defaults}}).}
}
\value{
\code{solve_range()} returns a list with the interim calculations
(\code{x_max}, \code{x_min}), the final value (\code{Range}), the solution
string (\code{solution}), and the bare formula (\code{formula}) in LaTeX
format. \code{range_formula} returns just the bare formula in LaTeX format
as a character string.
}
\description{
Range: calculate value (with rounding) and present solutions
}
\examples{
#' x <- sample(x = 1:10, size = 20, replace = TRUE)
solve_range(x, sub = "Y")

# Note that the raw data values get rounded by solve_range(x), so the exact
# value of the range will differ from the value calculated by range(x)
# when there are a lot of digits past zero.
(x <- rnorm(20))
Range <- solve_range(x, round_interim = 2)$Range
Range == diff(range(x))

# If you just want the bare formula as a string, use range_formula():
range_formula()

}
