% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{solve_s2}
\alias{solve_s2}
\alias{s2_formula}
\title{Sample Variance}
\usage{
solve_s2(
  x,
  SS,
  n,
  sub = "",
  sym = "X",
  SS.f = solve_sum_squares,
  abbrev_sum = 4,
  ...
)

s2_formula(sub = "", ...)
}
\arguments{
\item{x}{Numeric vector of raw data values.}

\item{SS}{Numeric scalar: sum of squares.}

\item{n}{Numeric scalar: sample size.}

\item{sub}{Character scalar. Name of numeric vector (reported as subscript in
solutions). Leave empty to report no subscript.}

\item{sym}{Character scalar. Symbol to represent x in formula (default: "X").
Only one character allowed. Only used when calculating SS from raw data
(\code{x}).}

\item{SS.f}{Formula to use for sum of squares calculation (either
solve_sum_squares or solve_sum_squares2). Only used when calculating SS
from raw data (\code{x}).}

\item{abbrev_sum}{Numeric scalar. Maximum length of x before it abbreviates
explicit summation within the solution using an ellipsis? (See
\code{\link{summation}}.) Only used when calculating SS from raw data
(\code{x}).}

\item{...}{Additional arguments to override default behaviors (see
\code{\link{handcalcs_defaults}}).}
}
\value{
\code{solve_s2()} returns a list with the interim values and
calculations (\code{x}, \code{SS}, \code{n}), the final value (\code{s2}),
the solution string (\code{solution}), and the bare formula
(\code{formula}) in LaTeX format. \code{s2_formula} returns just the bare
formula in LaTeX format as a character string.
}
\description{
Calculates sample variance (\eqn{s^2}) either from a set of raw values
(\code{x}) OR from the sum of squares (\code{SS}) and sample size (\code{n}).
If providing raw data, it will include the calculation of SS in the solution
string.
}
\examples{
# Simple usage: generate raw data, produce results
(x <- sample(x = 1:10, size = 20, replace = TRUE))
solve_s2(x = x, sub = 'X')

# Alternatively, provide SS and n
solve_s2(SS = 100, n = 10)

# Must provide enough values for calculation:
\dontrun{
solve_s2(SS = 100)
}

# If you just want the bare formula as a string, use a formula function():
s2_formula()

# Can set parameters to change symbols used:
s2_formula(sub = "Y")

}
