% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_one_sample.R
\name{solve_s_M}
\alias{solve_s_M}
\alias{s_M_formula}
\title{Standard Error of the Mean (Unknown Population Standard Deviation)}
\usage{
solve_s_M(SD, N, ...)

s_M_formula(...)
}
\arguments{
\item{SD}{Numeric scalar. Sample standard deviation.}

\item{N}{Numeric scalar. Sample size.}

\item{...}{Additional arguments to override default behaviors (see
\code{\link[=handcalcs_defaults]{handcalcs_defaults()}}}
}
\value{
\code{solve_s_M()} returns a list with the provided values (\code{SD}, \code{N}),
the interim calculations (\code{sqrtN}), the final value (\code{s_M}), the solution
string (\code{solution}), and the bare formula (\code{formula}) in LaTeX format.
\code{s_M_formula()} returns just the bare formula in LaTeX format as a
character string.
}
\description{
Calculates and provides solutions for the standard error of the mean (aka
standard deviation of the sampling distribution of the mean) using the sample
standard deviation (\eqn{s_{M} = s / \sqrt{N}}).
}
\examples{
solve_s_M(SD = 5, N = 25)
solve_s_M(SD = 5, N = 50)
solve_s_M(SD = 5, N = 100)

# If you just want the bare formula as a string, use a formula function():
s_M_formula()

}
