% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_paired_samples.R
\name{solve_s_M_D}
\alias{solve_s_M_D}
\alias{s_M_D_formula}
\title{Standard Error of the Mean Difference Score}
\usage{
solve_s_M_D(s_D = NULL, SD_D = NULL, n, ...)

s_M_D_formula(...)
}
\arguments{
\item{s_D, SD_D}{Numeric scalar. Standard deviation of difference scores. May
be named either \code{s_D} or \code{SD_D}.}

\item{n}{Numeric scalar. Sample size.}

\item{...}{Additional arguments to override default behaviors (see
\code{\link[=handcalcs_defaults]{handcalcs_defaults()}}}
}
\value{
\code{solve_s_M_D()} returns a list with the provided values (\code{s_D},
\code{SD_D}, \code{n}), the interim calculations (\code{sqrt_n}), the final value
(\code{s_M_D}), the solution string (\code{solution}), and the bare formula
(\code{formula}) in LaTeX format. \code{s_M_D_formula()} returns just the bare
formula in LaTeX format as a character string.
}
\description{
Calculates and provides solutions for the standard error of the mean
difference score using the standard deviation of the difference scores
(\eqn{s_{M_{D}} = s_{D} / \sqrt{n}}).
}
\examples{
# Starting with raw scores, calculate the difference scores first, then the SD:
x_1 <- rnorm(10)
x_2 <- rnorm(10)
x_D <- solve_diff_score(x_1 = x_1, x_2 = x_2)$x_D
s_D <- solve_sd_diff(x_D = x_D)$s_D
solve_s_M_D(s_D = s_D, n = 10)

solve_s_M_D(s_D = 5, n = 25)
solve_s_M_D(s_D = 5, n = 50)
solve_s_M_D(s_D = 5, n = 100)

# If you just want the bare formula as a string, use a formula function():
s_M_D_formula()

}
