% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_indep_samples.R
\name{solve_s_M_diff}
\alias{solve_s_M_diff}
\alias{s_M_diff_formula}
\title{Standard Error of Difference Between Two Independent Sample Means}
\usage{
solve_s_M_diff(s_p2, s_1, s_2, SD_1, SD_2, n_1, n_2, ...)

s_M_diff_formula(...)
}
\arguments{
\item{s_p2}{Numeric scalar. Pooled variance.}

\item{s_1, SD_1}{Numeric scalar. Standard deviation of sample 1.}

\item{s_2, SD_2}{Numeric scalar. Standard deviation of sample 2.}

\item{n_1}{Numeric scalar. Sample size of sample 1.}

\item{n_2}{Numeric scalar. Sample size of sample 2.}

\item{...}{Additional arguments to override default behaviors (see
\code{\link[=handcalcs_defaults]{handcalcs_defaults()}}}
}
\value{
\code{solve_s_M_diff()} returns a list with the provided values (\code{s_p2},
\code{s_1}, \code{s_2}, \code{SD_1}, \code{SD_2} \code{n_1}, \code{n_2}), the interim calculations (\code{s_p2_n1},
\code{s_p2_n2}, \code{s_M_diff2}), the final value (\code{s_M_diff}), the solution string
(\code{solution}), and the bare formula (\code{formula}) in LaTeX format (note: empty
values are removed from the list). \code{s_M_diff_formula()} returns just the
bare formula in LaTeX format as a character string.
}
\description{
Calculates the standard error of the difference between two independent
sample means (\code{s_M_diff}), used as the error term in the independent-samples
\emph{t}-test. Requires sample sizes (\code{n_1} and \code{n_2}) and either the pooled
variance (\code{s_p2}), or both samples' standard deviations (accepting either
\code{s_1} and \code{s_2} or \code{SD_1} and \code{SD_2}).
}
\details{
The standard error of the mean difference will be calculated using the pooled
variance as
\eqn{s_{M_{1}-M_{2}}=\sqrt{\frac{s_{p}^{2}}{n_{1}}+\frac{s_{p}^{2}}{n_{2}}}}.
If providing sample sizes and standard deviations instead, pooled variance
will be calculated from that information using the \verb{[solve_s_p2()]} function.
}
\examples{
solve_s_M_diff(s_p2 = 10, n_1 = 50, n_2 = 48)
solve_s_M_diff(s_1 = 4.5, s_2 = 5, n_1 = 50, n_2 = 48)

# If providing standard deviation, can use either s_1 and s_2, or SD_1 and SD_2
solve_s_M_diff(SD_1 = 4.5, SD_2 = 5, n_1 = 50, n_2 = 48)

}
\seealso{
\code{\link[=solve_s_p2]{solve_s_p2()}}, \code{\link[=solve_t_indep_samples]{solve_t_indep_samples()}}
}
