% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_indep_samples.R
\name{solve_s_p2}
\alias{solve_s_p2}
\alias{s_p2_formula}
\title{Pooled Variance}
\usage{
solve_s_p2(SS_1, SS_2, s_1, s_2, SD_1, SD_2, n_1, n_2, ...)

s_p2_formula(type = "SS", ...)
}
\arguments{
\item{SS_1}{Numeric scalar. Sum of squares of sample 1.}

\item{SS_2}{Numeric scalar. Sum of squares of sample 2.}

\item{s_1, SD_1}{Numeric scalar. Standard deviation of sample 1.}

\item{s_2, SD_2}{Numeric scalar. Standard deviation of sample 2.}

\item{n_1}{Numeric scalar. Sample size of sample 1.}

\item{n_2}{Numeric scalar. Sample size of sample 2.}

\item{...}{Additional arguments to override default behaviors (see
\code{\link[=handcalcs_defaults]{handcalcs_defaults()}}}
}
\value{
\code{solve_s_p2()} returns a list with the provided values (\code{SS_1},
\code{SS_2}, \code{s_1}, \code{s_2}, \code{SD_1}, \code{SD_2} \code{n_1}, \code{n_2}), the interim
calculations (\code{df_1}, \code{df_2}, \code{df} \code{s_12}, \code{s_22}, \code{df_s_12}, \code{df_s_22},
\code{SS}), the final value (\code{s_p2}), the solution string (\code{solution}), and the
bare formula (\code{formula}) in LaTeX format (note: empty values are removed
from the list). \code{s_p2_formula()} returns just the bare formula in LaTeX
format as a character string.
}
\description{
Calculates the pooled variance (\code{s_p2}) for two independent samples from
sample sizes (\code{n_1} and \code{n_2}) and either sums of squares for each sample
(\code{SS_1} and \code{SS_2}), or both samples' standard deviations (accepting either
\code{s_1} and \code{s_2} or \code{SD_1} and \code{SD_2}).
}
\details{
If providing sample sizes and sums of squares, pooled variance will be
calculated as \eqn{s_p^2=\frac{SS_1+SS_2}{n_1+n_2-2}}. If providing sample
sizes and standard deviations, pooled variance will be calculated as
\eqn{s_p^2=\frac{(n_{1}-1)(s_{1}^{2})+(n_{2}-1)(s_{2}^{2})}{n_1+n_2-2}}.
}
\examples{
solve_s_p2(SS_1 = 15, SS_2 = 20, n_1 = 40, n_2 = 50)
solve_s_p2(SD_1 = 5, SD_2 = 6, n_1 = 40, n_2 = 50)

}
\seealso{
\code{\link[=solve_s_M_diff]{solve_s_M_diff()}}, \code{\link[=solve_t_indep_samples]{solve_t_indep_samples()}}
}
