% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.R
\name{solve_sd}
\alias{solve_sd}
\alias{sd_formula}
\title{Sample Standard Deviation}
\usage{
solve_sd(
  x,
  SS,
  n,
  s2,
  SD2,
  sub_val = "",
  sym_x = "X",
  sub_x = "",
  SS.f = solve_sum_squares,
  ...
)

sd_formula(sub_val = "", ...)
}
\arguments{
\item{x}{Numeric vector of raw data values.}

\item{SS}{Numeric scalar: sum of squares.}

\item{n}{Numeric scalar: sample size.}

\item{s2, SD2}{Numeric scalar: sample variance (\eqn{s^2}). May name it
either as \code{s2} or \code{SD2}.}

\item{sub_val}{Character scalar. Subscript for the value to be calculated in
the formula, in this case the mean (e.g., \eqn{s_{x}}). Leave empty to
report no subscript.}

\item{sym_x}{Character scalar. Symbol to represent x in formula when
calculating from raw data (default: "X").}

\item{sub_x}{Character scalar. Subscript for x in the formula when
calculating from raw data (e.g., \eqn{X_{D}})}

\item{SS.f}{Formula to use for sum of squares calculation (either
solve_sum_squares or solve_sum_squares2). Only used when calculating SS
from raw data (\code{x}).}

\item{...}{Additional arguments to override default behaviors (see
\code{\link{handcalcs_defaults}}).}
}
\value{
\code{solve_sd()} returns a list with the interim values and
calculations (\code{x}, \code{SS}, \code{n}, \code{s2}), the final value (both as \code{s} and as
\code{SD}), the solution string (\code{solution}), and the bare formula (\code{formula})
in LaTeX format. \code{sigma2_formula} returns just the bare formula in LaTeX
format as a character string.
}
\description{
Calculates sample standard deviation (\eqn{s = \sqrt[SS / (n - 1)]}) either
from a set of raw values (\code{x}) OR from the sum of squares (\code{SS}) and sample
size (\code{n}), OR from the sample variance (\code{s2}). If providing raw data, it
will include the calculation of SS in the solution string. If providing \code{SS}
and \code{n}, it will include the calculation of the sample variance in the
solution string. Otherwise, if only \code{s2} is provided, will simply show the
final calculation (\eqn{\sqrt(s^2)}).
}
\examples{
# Simple usage: generate raw data, produce results
(x <- sample(x = 1:10, size = 20, replace = TRUE))
solve_s(x = x, sub_val = 'X')

# Alternatively, provide SS and n
solve_s(SS = 100, n = 10)

# Or, simply provide s2
solve_s(s2 = 10)

# Must provide enough values for calculation:
\dontrun{
solve_s(SS = 100)
}

# If you just want the bare formula as a string, use a formula function():
s_formula()

# Can set parameters to change symbols used:
s_formula(sub_val = "Y")
}
