% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.R
\name{solve_sigma}
\alias{solve_sigma}
\alias{sigma_formula}
\title{Population Standard Deviation}
\usage{
solve_sigma(
  x,
  SS,
  n,
  sigma2,
  sub_val = "",
  sym_x = "X",
  sub_x = "",
  SS.f = solve_sum_squares,
  ...
)

sigma_formula(sub_val = "", ...)
}
\arguments{
\item{x}{Numeric vector of raw data values.}

\item{SS}{Numeric scalar: sum of squares.}

\item{n}{Numeric scalar: sample size.}

\item{sigma2}{Numeric scalar: population variance (\eqn{sigma^2}).}

\item{sub_val}{Character scalar. Subscript for the value to be calculated in the
formula, in this case the population standard deviation (e.g.,
\eqn{\sigma_{x}}). Leave empty to report no subscript.}

\item{sym_x}{Character scalar. Symbol to represent x in formula when
calculating from raw data (default: "X").}

\item{sub_x}{Character scalar. Subscript for x in the formula when
calculating from raw data (e.g., \eqn{X_{D}})}

\item{SS.f}{Formula to use for sum of squares calculation (either
solve_sum_squares or solve_sum_squares2). Only used when calculating SS
from raw data (\code{x}).}

\item{...}{Additional arguments to override default behaviors (see
\code{\link{handcalcs_defaults}}).}
}
\value{
\code{solve_sigma2()} returns a list with the interim values and
calculations (\code{x}, \code{SS}, \code{n}, \code{sigma2}), the final
value (\code{sigma}), the solution string (\code{solution}), and the bare
formula (\code{formula}) in LaTeX format. \code{sigma_formula} returns just
the bare formula in LaTeX format as a character string.
}
\description{
Calculates population standard deviation (\eqn{\sigma = \sqrt(SS / n)})
either from a set of raw values (\code{x}) OR from the sum of squares
(\code{SS}) and sample size (\code{n}), OR from the population variance
(\code{sigma2}). If providing raw data, it will include the calculation of SS
in the solution string. If providing \code{SS} and \code{n}, it will include
the calculation of the population variance in the solution string. Otherwise,
if only \code{sigma2} is provided, will simply show the final calculation
(\eqn{\sqrt(\sigma^2)}).
}
\examples{
# Simple usage: generate raw data, produce results
(x <- sample(x = 1:10, size = 20, replace = TRUE))
solve_sigma(x = x, sub = 'X')

# Alternatively, provide SS and n
solve_sigma(SS = 100, n = 10)

# Or, simply provide sigma2
solve_sigma(sigma2 = 10)

# Must provide enough values for calculation:
\dontrun{
solve_sigma(SS = 100)
}

# If you just want the bare formula as a string, use a formula function():
sigma_formula()

# Can set parameters to change symbols used:
sigma_formula(sub_val = "Y")
}
