% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_one_sample.R
\name{solve_sigma_M}
\alias{solve_sigma_M}
\alias{sigma_M_formula}
\title{Standard Error of the Mean (Known Population Standard Deviation)}
\usage{
solve_sigma_M(sigma, n, ...)

sigma_M_formula(...)
}
\arguments{
\item{sigma}{Numeric scalar. Population standard deviation.}

\item{n}{Numeric scalar. Sample size.}

\item{...}{Additional arguments to override default behaviors (see
\code{\link[=handcalcs_defaults]{handcalcs_defaults()}}}
}
\value{
\code{solve_sigma_M()} returns a list with the provided values (\code{sigma},
\code{n}), the interim calculations (\code{sqrt_n}), the final value (\code{sigma_M}), the
solution string (\code{solution}), and the bare formula (\code{formula}) in LaTeX
format. \code{sigma_M_formula()} returns just the bare formula in LaTeX format
as a character string.
}
\description{
Calculates and provides solutions for the standard error of the mean (aka
standard deviation of the sampling distribution of the mean) when the
population standard deviation (\eqn{\sigma}) is known (\eqn{\sigma_{M} = \sigma
/ \sqrt (n)}).
}
\examples{
solve_sigma_M(sigma = 5, n = 25)
solve_sigma_M(sigma = 5, n = 50)
solve_sigma_M(sigma = 5, n = 100)

# If you just want the bare formula as a string, use a formula function():
sigma_M_formula()

}
