% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_squares.R
\name{solve_sum_squares}
\alias{solve_sum_squares}
\alias{sum_squares_formula}
\alias{sum_squares_table}
\alias{solve_sum_squares2}
\alias{sum_squares_formula2}
\alias{sum_squares_table2}
\title{Sum of Squares}
\usage{
solve_sum_squares(x, sub = "", sym = "X", ...)

sum_squares_formula(sub = "", sym = "X", ...)

sum_squares_table(SS.lst, sub = "", sym = "X", ...)

solve_sum_squares2(x, sub = "", sym = "X", ...)

sum_squares_formula2(sub = "", sym = "X", ...)

sum_squares_table2(SS.lst, sym = "X", ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{sub}{Character scalar. Name of numeric vector (reported as subscript in
solutions). Leave empty to report no subscript.}

\item{sym}{Character scalar. Symbol to represent x in formula (default: "X").
Only one character allowed.}

\item{...}{Additional arguments to override default behaviors (see
\code{\link{handcalcs_defaults}}).}

\item{SS.lst}{List of values produced by \code{solve_sum_squares} or
\code{solve_sum_squares2}. Be sure to use the table function that
corresponds to the solution function.}
}
\value{
The two solutions functions (\code{solve_sum_squares},
\code{solve_sum_squares2}) return a named list with the interim
calculations, the final value (\code{SS}), the solution string
(\code{solution}), and the bare formula (\code{formula}) in LaTeX format.
(Note that the two functions will return different values for the interim
calculations based on using different formulas.) The two formula functions
(\code{sum_squares_formula}, \code{sum_squares_formula2}) return the bare
formula in LaTeX format as a character string. The two table functions
(\code{sum_squares_table}, \code{sum_squares_table2}) return a gt object
which can be rendered into HTML or PDF.
}
\description{
Calculate value of sum of squares (with rounding) and present solutions.
\code{solve_sum_squares()} produces the interim calculations and a full
solution string, starting with the bare formula. If you just want the bare
formula, use \code{sum_squares_formula()}. To generate the table
demonstrating the calculations, use \code{sum_squares_table}.
}
\details{
The standard functions use the conceptual formula: \deqn{SS = \sum(X - M)^2}.

The '2' functions (\code{solve_sum_squares2}, \code{sum_squares_formula2})
use the alternative computational formula: \deqn{SS = \sum(X^2) - [(\sum
X)^2] / n}
}
\examples{
# The formulas will return similar results when rounding is minimal
(x <- sample(x = 1:10, size = 20, replace = TRUE))
solve_sum_squares(x)$SS
solve_sum_squares2(x)$SS

# They are more likely to differ when there is a lot of rounding:
(x <- rnorm(20))
solve_sum_squares(x)
solve_sum_squares2(x)

# Can override the default values for rounding interim and final values:
solve_sum_squares(x, round_interim = 2, round_final = 2)

# The value of sub can be long, but not sym.
\dontrun{
solve_sum_squares(x, sub = "Donuts", sym = "Donuts")
}

# If you just want the bare formula as a string, use a formula function():
sum_squares_formula()
sum_squares_formula2()

# Can set parameters to change symbols used:
sum_squares_formula(sub = "Y", sym = "Y")

}
