% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_scores.R
\name{solve_sum_squares_diff}
\alias{solve_sum_squares_diff}
\alias{solve_sum_squares2_diff}
\alias{solve_s2_diff}
\alias{solve_sd_diff}
\title{Difference Scores: Sum of Squares, Variance, Standard Deviation}
\usage{
solve_sum_squares_diff(x_D, sym_x = "X", ...)

solve_sum_squares2_diff(x_D, sym_x = "X", ...)

solve_s2_diff(x_D, SS_D, n, sym_x = "X", SS.f = solve_sum_squares, ...)

solve_sd_diff(x_D, SS_D, n, s2_D, sym_x = "X", SS.f = solve_sum_squares, ...)
}
\arguments{
\item{x_D}{Numeric vector of difference scores.}

\item{sym_x}{Character scalar. Symbol to represent x in formula (default:
"X").}

\item{...}{Additional arguments to override default behaviors (see
\verb{[handcalcs_defaults()]}).}

\item{SS_D}{Numeric scalar: sum of squares of difference scores.}

\item{n}{Numeric scalar: sample size.}

\item{SS.f}{Formula to use for sum of squares calculation (either
'solve_sum_squares' or 'solve_sum_squares2'). Only used when calculating SS
from raw difference scores (\code{x_D}).}

\item{s2_D}{Numeric scalar: sample variance of difference scores (\eqn{s_{D}^2}).}
}
\value{
List of values. See the original functions for more information:
\code{\link[=solve_sum_squares]{solve_sum_squares()}}, \code{\link[=solve_sum_squares2]{solve_sum_squares2()}}, \code{\link[=solve_s2]{solve_s2()}},
\code{\link[=solve_sd]{solve_sd()}}.
}
\description{
Convenience wrappers for calculating sum of squares (\code{\link[=solve_sum_squares]{solve_sum_squares()}},
\code{\link[=solve_sum_squares2]{solve_sum_squares2()}}), variance (\code{\link[=solve_s2]{solve_s2()}}), and standard deviation
(\code{\link[=solve_sd]{solve_sd()}}) when dealing with difference scores. Essentially just
ensures that they use the "D" subscript appropriately.
}
\examples{
solve_sum_squares_diff(x_D = 1:10)

solve_sum_squares2_diff(x_D = 1:10)

solve_s2_diff(x_D = 1:10)

solve_sd_diff(x_D = 1:10)
}
