% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical.R
\name{solve_t_crit}
\alias{solve_t_crit}
\title{Critical values of \emph{t}}
\usage{
solve_t_crit(alpha = 0.05, df, direction = c("pos", "neg", "both"), ...)
}
\arguments{
\item{alpha}{Numeric scalar. Value of alpha (probability of a Type I error).
Defaults to 0.05.}

\item{df}{Numeric scalar. Degrees of freedom.}

\item{direction}{Which tail of the distribution should the area equivalent to
\code{alpha} be located: pos (t_obt > t_crit), neg (t_obt < t_crit), both
(|t_obt| > |t_crit|).}

\item{...}{Additional arguments to override default behaviors (see
\code{\link[=handcalcs_defaults]{handcalcs_defaults()}}).}
}
\value{
Returns a list with the given values (\code{alpha}, \code{df}, \code{direction}),
the final value (\code{t_crit}), and the solution string (\code{solution}).
}
\description{
Calculates the critical value of \emph{t} for a given \eqn{\alpha}, degrees of
freedom (\code{df}), and the direction of the hypothesis, along with the solution
string in LaTeX format.
}
\details{
Note that the returned critical value of \emph{t} (\code{t_crit}) is rounded to the value
of \code{round_t} instead of the value of \code{round_interim} or \code{round_final} (see
\code{\link[=handcalcs_defaults]{handcalcs_defaults()}}).
}
\examples{

solve_t_crit(alpha = 0.01, df = 100, direction = 'pos')
solve_t_crit(alpha = 0.01, df = 100, direction = 'neg')
solve_t_crit(alpha = 0.01, df = 100, direction = 'both')

# In some cases, you may wish to change the default value for round_final:
solve_t_crit(alpha = 0.05, df = 100, direction = 'pos', round_final = 3)

}
