% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_indep_samples.R
\name{solve_t_indep_samples}
\alias{solve_t_indep_samples}
\alias{t_indep_samples_formula}
\title{Independent-samples \emph{t}-test}
\usage{
solve_t_indep_samples(
  M_1,
  M_2,
  s_M_diff,
  s_p2,
  s_1,
  s_2,
  SD_1,
  SD_2,
  n_1,
  n_2,
  ...
)

t_indep_samples_formula(...)
}
\arguments{
\item{M_1}{Numeric scalar. Mean of sample 1.}

\item{M_2}{Numeric scalar. Mean of sample 2.}

\item{s_M_diff}{Numeric scalar. Standard error of the difference between two
independent sample means. If not provided, can be calculated using the
other arguments using the \verb{[solve_s_M_diff()]} function.}

\item{s_p2}{Numeric scalar. Pooled variance.}

\item{s_1, SD_1}{Numeric scalar. Standard deviation of sample 1.}

\item{s_2, SD_2}{Numeric scalar. Standard deviation of sample 2.}

\item{n_1}{Numeric scalar. Sample size of sample 1.}

\item{n_2}{Numeric scalar. Sample size of sample 2.}

\item{...}{Additional arguments to override default behaviors (see
\code{\link[=handcalcs_defaults]{handcalcs_defaults()}}}
}
\value{
\code{solve_t_indep_samples()} returns a list with the provided values
(\code{M_1}, \code{M_2}, \code{s_M_diff}, \code{s_p2}, \code{s_1}, \code{s_2}, \code{SD_1}, \code{SD_2} \code{n_1},
\code{n_2}), the interim calculations (\code{M_diff}), the final value (\code{t_obt}), the
solution string (\code{solution}), and the bare formula (\code{formula}) in LaTeX
format (note: empty values are removed from the list).
\code{t_indep_samples_formula()} returns just the bare formula in LaTeX format
as a character string.
}
\description{
Calculates and provides solutions for the independent-samples \emph{t}-test (i.e.,
\code{t_obt}), comparing the calculated difference between two independent sample
means (\eqn{M_{1} - M_{2}}) against the predicted difference under the null
hypothesis (\eqn{\mu_{1} - \mu_{2}}, which is always zero, and thus not an
argument to this function).
}
\details{
The error term (\eqn{s_{M_{1} - M_{2}}}) can be provided directly (as
\code{s_M_diff}) along with the two sample means (\code{M_1} and \code{M_2}) to calculate
\code{t_obt} as (\eqn{t_{obt}=\frac{M_{1}-M_{2}}{s_{M_{1}-M_{2}}}}). If the
standard error is not provided, it can be calculated using either the pooled
variance (\code{s_p2}) and the sample sizes (\code{n_1} and \code{n_2}), or both samples'
standard deviations (\code{s_1} and \code{s_2} or \code{SD_1} and \code{SD_2}) and sample sizes
(\code{n_1} and \code{n_2}) using the \verb{[solve_s_M_diff()]} function. In that case,
those calculations will be included in the solution string.

Note that \code{t_obt} values are rounded to the value of \code{round_t} instead of the
value of \code{round_final} (see \code{\link[=handcalcs_defaults]{handcalcs_defaults()}}).
}
\examples{

solve_t_indep_samples(M_1 = 5, M_2 = 4, s_M_diff = 0.5)

# If not providing the error term, it can be calculated either from the
# sample sizes and either the pooled variance or sample standard deviations
solve_t_indep_samples(M_1 = 5, M_2 = 4, s_p2 = 5, n_1 = 40, n_2 = 45)
solve_t_indep_samples(M_1 = 5, M_2 = 4, s_1 = 2, s_2 = 3, n_1 = 40, n_2 = 45)

# Standard deviations can be provided either as s_1 and s_2 or SD_1 and SD_2
solve_t_indep_samples(M_1 = 5, M_2 = 4, s_1 = 2, s_2 = 3, n_1 = 40, n_2 = 45)

}
