% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_one_sample.R
\name{solve_t_one_sample}
\alias{solve_t_one_sample}
\alias{t_one_sample_formula}
\title{One-sample \emph{t}-test}
\usage{
solve_t_one_sample(M, mu, s, SD, n, s_M, ...)

t_one_sample_formula(...)
}
\arguments{
\item{M}{Numeric scalar. Sample mean.}

\item{mu}{Numeric scalar. Population mean.}

\item{s, SD}{Numeric scalar. Sample standard deviation. Required if s_M is not
provided. May be named either \code{s} or \code{SD}.}

\item{n}{Numeric scalar. Sample size. Required if s_M is not provided.}

\item{s_M}{Numeric scalar. Standard error of the mean. If not provided, will
be calculated from \code{s} and \code{n} (using \code{\link[=solve_s_M]{solve_s_M()}}) and included in the
solution string.}

\item{...}{Additional arguments to override default behaviors (see
\code{\link[=handcalcs_defaults]{handcalcs_defaults()}}}
}
\value{
\code{solve_t_one_sample()} returns a list with the provided values (\code{M},
\code{mu}, \code{s}, \code{SD}, \code{n}), the interim calculations (\code{df}, \code{s_M}, \code{M_diff}), the
final value (\code{t_obt}), the solution string (\code{solution}), and the bare
formula (\code{formula}) in LaTeX format. \code{t_one_sample_formula()} returns just
the bare formula in LaTeX format as a character string.
}
\description{
Calculates and provides solutions for the one-sample \emph{t}-test (i.e., \code{t_obt}),
comparing an obtained sample mean against the population mean (\eqn{\mu}) when
the population standard deviation (\eqn{\sigma}) must be estimated using the
sample standard deviation (\eqn{s}, aka \code{SD}).
}
\details{
Note that \code{t_obt} values are rounded to the value of \code{round_t} instead of the
value of \code{round_final} (see \code{\link[=handcalcs_defaults]{handcalcs_defaults()}}).
}
\examples{
solve_t_one_sample(M = 6, mu = 5, SD = 2, n = 100)

# If you just want the formula:
t_one_sample_formula()

}
