% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centraltendency.R
\name{solve_weighted_mean}
\alias{solve_weighted_mean}
\alias{weighted_mean_formula}
\title{Weighted (aka overall) Mean}
\usage{
solve_weighted_mean(Samples, ...)

weighted_mean_formula(...)
}
\arguments{
\item{Samples}{List of means and sample sizes. Must be a list of vectors with
names M and n, one vector per subgroup (see examples).}

\item{...}{Additional arguments to override default behaviors (see
\code{\link{handcalcs_defaults}}).}
}
\value{
\code{solve_weighted_mean()} returns a list with vectors for the
means (\code{M}) and sample sizes (\code{n}), the interim calculations (\code{Sum_M_x_n},
\code{Sum_n}), the final value (\code{M_w}), the solution string (\code{solution}), and
the bare formula (\code{formula}) in LaTeX format. \code{weighted_mean_formula}
returns just the bare formula in LaTeX format as a character string.
}
\description{
Calculate value of weighted (aka overall) mean (with rounding) and present
solutions. \code{solve_weighted_mean()} produces the interim calculations and
a full solution string, starting with the bare formula. If you just want the
bare formula, use \code{weighted_mean_formula()}.
}
\examples{
l <- list(c(M = 10, n = 20), c(M = 15, n = 25), c(M = 20, n = 30), c(M = 25, n = 35))
solve_weighted_mean(l)

# For a large number of subgroups, you may wish to truncate the solution:
solve_weighted_mean(l, abbrev_sum = 3)

# If you just want the bare formula as a string, use weighted_mean_formula():
weighted_mean_formula()
}
