% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_scores.R
\name{solve_x_to_z}
\alias{solve_x_to_z}
\alias{x_to_z_formula}
\alias{solve_z_to_x}
\alias{z_to_x_formula}
\title{Converting Raw Scores to Z-scores}
\usage{
solve_x_to_z(x, M, SD, sub_val = "", sym_x = "x", population = FALSE, ...)

x_to_z_formula(sub_val = "", sym_x = "x", population = FALSE, ...)

solve_z_to_x(z, M, SD, sub_val = "", sym_x = "x", population = FALSE, ...)

z_to_x_formula(sub_val = "", sym_x = "x", population = FALSE, ...)
}
\arguments{
\item{x}{Numeric scalar. Raw score to convert to a z-score.}

\item{M}{Numeric scalar. Mean of distribution.}

\item{SD}{Numeric scalar. Standard Deviation of distribution.}

\item{sub_val}{Character scalar. Subscript for the value to be calculated in the
formula, in this case the \emph{z}-score (e.g., \eqn{z_{x}}). Leave empty to report
no subscript.}

\item{sym_x}{Character scalar. Symbol to represent x in formula (default:
"X").}

\item{population}{Logical. Is this a population, rather than a sample? This
is only relevant for displaying the formula in the solutions, which will
use Greek letters to denote the population mean (\eqn{\mu}) and standard
deviation (\eqn{\sigma}).}

\item{...}{Additional arguments to override default behaviors (see
\code{\link{handcalcs_defaults}}).}

\item{z}{Numeric scalar. Z-score to convert to a raw score (x).}
}
\value{
\code{solve_x_to_z()} returns a list with the given values, the
interim calculations (\code{M_diff}), the final value (\code{z}), the
solution string (\code{solution}), and the bare formula (\code{formula}) in
LaTeX format.

\code{solve_z_to_x()} returns a list with the given values, the interim
calculations (\code{zSD}), the final value (\code{x}), the solution string
(\code{solution}), and the bare formula (\code{formula}) in LaTeX format.

The two formula functions (\code{x_to_z_formula}, \code{z_to_x_formula})
return just the bare formula in LaTeX format as a character string.
}
\description{
Functions to convert raw scores to z-scores and vice versa.
\code{solve_x_to_z()} takes a raw score (\code{x}), along with the distribution's
mean (\code{M}) and standard deviation (\code{SD}), and calculates the standard score
(\code{z}). \code{solve_z_to_x()} takes a z-score (\code{z}), along with the
distribution's mean (\code{M}) and standard deviation (\code{SD}), and calculates the
raw score (\code{x}). If you just want the bare formula for either step, use
\code{x_to_z_formula()} or \code{z_to_x_formula()}.
}
\details{
Note that \emph{z}-scores are rounded to the value of \code{round_z} instead of the
value of \code{round_interim} or \code{round_final} (see \code{\link[=handcalcs_defaults]{handcalcs_defaults()}}).
}
\examples{
# Convert a raw score to a z-score:
M <- 10
SD <- 2
(z.lst <- solve_x_to_z(x = 5, M = M, SD = SD))

# And convert it back to a x
solve_z_to_x(z = z.lst$z, M = M, SD = SD)

# Or calculate a novel value:
solve_z_to_x(z = 1.1, M = M, SD = SD)

# Can change the symbols used:
# If you just want the bare formula as a string, use the formula functions:
x_to_z_formula()
z_to_x_formula()

# Set population to TRUE to show Greek symbols:
x_to_z_formula(population = TRUE)
z_to_x_formula(population = TRUE)

}
