% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical.R
\name{solve_z_crit}
\alias{solve_z_crit}
\title{Critical values of Z}
\usage{
solve_z_crit(alpha = 0.05, direction = c("pos", "neg", "both"), ...)
}
\arguments{
\item{alpha}{Value of alpha (probability of a Type I error). Defaults to
0.05.}

\item{direction}{Which tail of the distribution should the area equivalent to
\code{alpha} be located: pos (z_obt > z_crit), neg (z_obt < z_crit), both
(|z_obt| > |z_crit|).}

\item{...}{Additional arguments to override default behaviors (see
\code{\link[=handcalcs_defaults]{handcalcs_defaults()}}).}
}
\value{
Returns a list with the given values (\code{alpha}, \code{direction}), the
final value (\code{z_crit}), and the solution string (\code{solution}).
}
\description{
Calculates the critical value of \emph{z} for a given \eqn{\alpha} and the
direction of the hypothesis, along with the solution string in LaTeX format.
}
\details{
Note that the returned critical value of z (\code{z_crit}) is rounded to the value
of \code{round_z} instead of the value of \code{round_interim} or \code{round_final} (see
\code{\link[=handcalcs_defaults]{handcalcs_defaults()}}).
}
\examples{

solve_z_crit(alpha = 0.01, direction = 'pos')
solve_z_crit(alpha = 0.01, direction = 'neg')
solve_z_crit(alpha = 0.01, direction = 'both')

# In some cases, you may wish to change the default value for round_z:
solve_z_crit(alpha = 0.05, direction = 'pos', round_z = 3)

}
