% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{solve_z_to_auc}
\alias{solve_z_to_auc}
\title{Calculating Area Under the Normal Curve}
\usage{
solve_z_to_auc(z, type = c("above", "below", "between", "tails"), ...)
}
\arguments{
\item{z}{Numeric. For type values of \code{above} and \code{below}, must be a single
value. For type values of \code{between} and \code{tails}, can be of length 1 or 2.}

\item{type}{Character. Determines what type of area is to be calculated:
\code{above}, \code{below}, \code{between}, or \code{tails}.}

\item{...}{Additional arguments to override default behaviors (see
\code{\link[=handcalcs_defaults]{handcalcs_defaults()}}).}
}
\value{
Returns a list with the given value (\code{z}), the interim
calculations (\code{a1} and \code{a2} for \code{type = between} and \code{type = tails}), the
final value (\code{a}), and the solution string (\code{solution}).
}
\description{
Takes a given value of \code{z} (e.g., a \emph{z}-score) and calculates the specified
area under the curve (AUC). The type argument determines what area is being
calculated and how the value of \code{z} is interpreted:
}
\details{
\describe{

\item{\code{above}}{Calculates the area above the given value of \code{z}.}

\item{\code{below}}{Calculates the area above the given value of \code{z}.}

\item{\code{between}}{Calculates the area under the curve between two values
of \code{z}. When a single value of \code{z} is provided, calculates the area under the
curve between the positive and negative equivalent values of \code{z}.}

\item{\code{tails}}{Calculates the area under the curve more extreme (i.e.,
in the tails) of the two values of \code{z}. Note that when two values are
provided, one must be positive and one negative. When a single value of \code{z}
is provided, calculates the area under the curve more extreme than the
positive and negative equivalent values of \code{z} (in the tails).}

}

Note that \emph{z}-scores are rounded to the value of \code{round_z} instead of the
value of \code{round_interim} or \code{round_final} (see \code{\link[=handcalcs_defaults]{handcalcs_defaults()}}).
}
\examples{

# Calculates the area above and below a single z value.
solve_z_to_auc(1, 'above')
solve_z_to_auc(1, 'below')

# When two values are provided (for between and tails), can calculate area between:
solve_z_to_auc(z = c(-2, 1), 'between')
solve_z_to_auc(z = c(-2, 1), 'tails')

# When only one value of z is provided for between and tails, presumes symmetry:
solve_z_to_auc(2, 'between')
solve_z_to_auc(2, 'tails')

}
