% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_helpers.R
\name{summation}
\alias{summation}
\title{Explicit Summation}
\usage{
summation(sum_values, abbrev_sum = 5)
}
\arguments{
\item{sum_values}{Character vector (or vector of values that can be coerced
to character) of values to be summed.}

\item{abbrev_sum}{Numeric scalar indicating when to abbreviate the summation
(with an ellipsis). If the length of sum_values exceeds the value of
abbrev_sum, will return abbreviated string. (Note: minimum value is 3.)}
}
\value{
Explicit summation string (scalar). For instance,
\code{summation(1:4)} would return "1 + 2 + 3 + 4".
}
\description{
Takes a character vector of values that would be part of a summation sequence
(e.g., produced by \code{glue::glue()}) and returns a string enumerating the
process. For instance, if you wished to sum the integers 1 to 3, this would
be spelled out as "1 + 2 + 3". By default, will abbreviate the summation for
vectors longer than \verb{\code{abbrev_sum}} by using an ellipsis (LaTeX: \verb{\\cdots}).
}
\examples{
# Simplest use will take any string (or values that can be coerced):
summation(1:4)

# More complex uses are possible (e.g., sum of squared deviations)
x <- sample(1:20, size = 10)
M <- round(mean(x), 2)
sum_values <- glue::glue("(<<x>> - <<M>>)^2", .open = "<<", .close = ">>")
summation(sum_values)

# If you wish to avoid abbreviation, set the value higher:
summation(sum_values, abbrev_sum = length(sum_values))

}
