% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_digits}
\alias{get_digits}
\title{Get Minimum Digits}
\usage{
get_digits(x, max_digits = 4)
}
\arguments{
\item{x}{Numeric vector.}

\item{max_digits}{Numeric scalar: maximum number of digits.}
}
\value{
Numeric scalar: smallest number of digits to which the number(s) can
be rounded (up to max_digits) without losing precision.
}
\description{
Figures out the minimum number of digits (up to four) to report after the
decimal place.
}
\examples{

get_digits(rnorm(20))


}
