% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{get_freq_tbl}
\alias{get_freq_tbl}
\title{Frequency Tables}
\usage{
get_freq_tbl(
  x,
  grouped = TRUE,
  start = 0,
  width = 5,
  rf = TRUE,
  cf = FALSE,
  cp = FALSE,
  pr = FALSE,
  font_size = 12,
  ...
)
}
\arguments{
\item{x}{Numeric vector.}

\item{grouped}{Logical. Should this be a grouped frequency table?}

\item{start}{Numeric scalar. For ungrouped frequency tables, lowest value
represented in the table (default = \code{min(x)}). For grouped frequency
tables, the lower bound of the lowest interval.}

\item{width}{Numeric scalar. Width of interval for grouped frequency tables.
(Ignored if grouped is FALSE.)}

\item{rf}{Logical. Should a column for relative frequency be included
(default = TRUE)? Note: values will be rounded to the value of
round_interim.}

\item{cf}{Logical. Should a column for cumulative frequency be included
(default = FALSE)?}

\item{cp}{Logical. Should a column for cumulative percentage be included
(default = FALSE)? Note: values will be rounded to the value of
round_final.}

\item{pr}{Logical. Should a column for percentile rank be included (default =
FALSE)? Note: values will be rounded to the value of round_final.}

\item{font_size}{Numeric scalar. Font size of table in points (as rendered by
\verb{[gt::gt()]}).}

\item{...}{Additional arguments to override default behaviors (see
\code{\link[=handcalcs_defaults]{handcalcs_defaults()}}).}
}
\value{
A gt object.
}
\description{
Produces both ungrouped and grouped frequency tables (as a \verb{[gt::gt()]}
object) from a set of raw data.
}
\examples{

x <- sample(x = 1:20, size = 20, replace = TRUE)

# Ungrouped:
get_freq_tbl(x, grouped = FALSE)
get_freq_tbl(x, grouped = FALSE, cf = TRUE, cp = TRUE, pr = TRUE, round_final = 2)

# Grouped:
get_freq_tbl(x, grouped = TRUE)
get_freq_tbl(x, grouped = TRUE, width = 2, cf = TRUE, cp = TRUE, pr = TRUE, round_final = 2)

}
