% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{handcalcs_defaults}
\alias{handcalcs_defaults}
\title{Default behaviors for handcalcs functions}
\usage{
handcalcs_defaults()
}
\value{
Named list of default values
}
\description{
Sets default behaviors for handcalcs functions related to rounding for
interim/final calculations and handling of LaTeX solution string.
}
\details{
The default behavior creates a self-contained LaTeX solution string (enclosed
in an aligned display math environment). If you are writing the LaTeX code
manually, (e.g., if you wish to present multiple equations within the same
environment), you can provide the named value as an argument to any solution
function. If you wish to change the global defaults (applied to all
functions), set the corresponding value of the named list ("handcalcs") in
\code{options()}.

\describe{

\item{\code{round_to}}{Round based on the number of digits past the decimal
('decimal'), or based on the number of significant figures ('sigfig'). Note:
this parameter can only be set as a global default, not within individual
functions.}

\item{\code{round_interim}}{Number of digits to round for interim
calculations (default = 4). Note: It is strongly recommended that you change
this value if you are using sigfigs instead of decimals for rounding.}

\item{\code{round_final}}{Number of digits to round for final result (default
= 4). Note: It is strongly recommended that you change this value if you are
using sigfigs instead of decimals for rounding.}

\item{\code{round_z}}{Number of digits to round values of \code{z}, such as for
\emph{z}-scores (default = 2). This is primarily useful to ensure that the \emph{z}
values being reported conform to the values presented in a z-table. Note that
this overrides \code{round_interim} and \code{round_final}.}

\item{\code{add_math}}{Enclose solution string in LaTeX math block (\verb{$$};
default = TRUE)? If set to \code{FALSE}, must manually include the math block tags
in order for the solution to render properly.}

\item{\code{add_aligned}}{Enclose solution in LaTeX aligned block
(\verb{\\begin\{aligned\}}) (default = TRUE)? If set to \code{FALSE}, allows for multiple
solutions to be set in the same aligned environment.}

\item{\code{use_aligned}}{Use "&=" instead of "=" in solution string to
generate aligned equations (default = TRUE).}

}
}
\examples{
handcalcs_defaults()

}
