% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centraltendency.R
\name{solve_mean}
\alias{solve_mean}
\alias{mean_formula}
\title{Arithmetic Mean}
\usage{
solve_mean(x, sub_val = "", sym_x = "X", sub_x = "", ...)

mean_formula(sub_val = "", sym_x = "X", sub_x = "", ...)
}
\arguments{
\item{x}{Numeric vector of values for which to calculate the mean.}

\item{sub_val}{Character scalar. Subscript for the value to be calculated in
the formula, in this case the mean (e.g., \eqn{M_{x}}). Leave empty to
report no subscript.}

\item{sym_x}{Character scalar. Symbol to represent x in formula (default:
"X").}

\item{sub_x}{Character scalar. Subscript for x in the formula (e.g.,
\eqn{X_{D}})}

\item{...}{Additional arguments to override default behaviors (see
\code{\link{handcalcs_defaults}}).}
}
\value{
\code{solve_mean()} returns a list with the interim calculations
(\code{SumX}, \code{n}), the final value (\code{M}), the solution string (\code{solution}), and
the bare formula (\code{formula}) in LaTeX format. \code{mean_formula} returns just
the bare formula in LaTeX format as a character string.
}
\description{
Calculate value of arithmetic mean (with rounding) and present solutions.
\code{solve_mean()} produces the interim calculations and a full solution
string, starting with the bare formula. If you just want the bare formula,
use \code{mean_formula()}.
}
\examples{
x <- sample(x = 1:10, size = 20, replace = TRUE)
solve_mean(x)

# Note that the value calculated by solve_mean(x) will be different than the
# value calculated by mean(x) when rounding occurs.
(x <- rnorm(20))
solve_mean(x, sub = "Y", sym = "Y")

# Can override the default values for rounding interim and final values:
(x <- rnorm(20))
solve_mean(x, round_interim = 2, round_final = 2)

# The value of sub can be long, but not sym.
\dontrun{
solve_mean(x, sub = "Donuts", sym = "Donuts")
}

# If you just want the bare formula as a string, use mean_formula():
mean_formula()

# Can set parameters to change symbols used:
mean_formula(sub = "Y", sym = "Y")

}
