% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centraltendency.R
\name{solve_mean}
\alias{solve_mean}
\title{Solve arithmetic mean}
\usage{
solve_mean(x, nm = "X", precision = 4, digits = 4)
}
\arguments{
\item{x}{A numeric vector}

\item{nm}{Name of numeric vector (reported as subscript in solutions)}

\item{precision}{Number of digits to round for interim calculation steps}

\item{digits}{Number of digits to report for final result}
}
\value{
A list with the interim calculations (SumX, N), the final value (M), and the solution (slv)
}
\description{
Solve arithmetic mean
}
\examples{
x <- 1:10
solve_mean(x)

(x<- rnorm(20))
solve_mean(x, nm='Y')


}
