% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centraltendency.R
\name{solve_mode}
\alias{solve_mode}
\title{Mode: calculate value (with rounding) and present solutions}
\usage{
solve_mode(x, sub_val = "", ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{sub_val}{Character scalar. Subscript for the value to be calculated in the
formula, in this case the mode (e.g., \eqn{Mode_{x}}). Leave empty to
report no subscript.}

\item{...}{Additional arguments to override default behaviors (see
\code{\link{handcalcs_defaults}}).}
}
\value{
A list with the mode value(s) (\code{Mode}) and the solution
(\code{solution}) in LaTeX format.
}
\description{
Mode: calculate value (with rounding) and present solutions
}
\examples{
#' x <- sample(x = 1:10, size = 20, replace = TRUE)
solve_mode(x)

# Note that the raw data values get rounded by solve_mode(x), so the exact
# value(s) provided may differ from the values in the (unrounded) raw data.
(x <- rnorm(20))
solve_mode(x, round_interim = 2)$Mode

}
