% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.R
\name{solve_range}
\alias{solve_range}
\alias{range_formula}
\title{Range: calculate value (with rounding) and present solutions}
\usage{
solve_range(x, sub_val = "", sym_x = "X", ...)

range_formula(sub_val = "", sym_x = "X", ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{sub_val}{Character scalar. Subscript for the value to be calculated the
formula, in this case the range (e.g., \eqn{Range_{x}}). Leave empty to report
no subscript.}

\item{sym_x}{Character scalar. Symbol to represent x in formula (default:
"X").}

\item{...}{Additional arguments to override default behaviors (see
\code{\link{handcalcs_defaults}}).}
}
\value{
\code{solve_range()} returns a list with the interim calculations
(\code{x_max}, \code{x_min}), the final value (\code{Range}), the solution
string (\code{solution}), and the bare formula (\code{formula}) in LaTeX
format. \code{range_formula} returns just the bare formula in LaTeX format
as a character string.
}
\description{
Range: calculate value (with rounding) and present solutions
}
\examples{
#' x <- sample(x = 1:10, size = 20, replace = TRUE)
solve_range(x, sub = "Y")

# Note that the raw data values get rounded by solve_range(x), so the exact
# value of the range will differ from the value calculated by range(x)
# when there are a lot of digits past zero.
(x <- rnorm(20))
Range <- solve_range(x, round_interim = 2)$Range
Range == diff(range(x))

# If you just want the bare formula as a string, use range_formula():
range_formula()

}
