% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.R
\name{solve_sd}
\alias{solve_sd}
\alias{sd_formula}
\title{Sample Standard Deviation}
\usage{
solve_sd(
  x,
  SS,
  n,
  s2,
  sub = "",
  sym = "X",
  SS.f = solve_sum_squares,
  abbrev_sum = 4,
  ...
)

sd_formula(sub = "", ...)
}
\arguments{
\item{x}{Numeric vector of raw data values.}

\item{SS}{Numeric scalar: sum of squares.}

\item{n}{Numeric scalar: sample size.}

\item{s2}{Numeric scalar: sample variance (\eqn{s^2}).}

\item{sub}{Character scalar. Name of numeric vector (reported as subscript in
solutions). Leave empty to report no subscript.}

\item{sym}{Character scalar. Symbol to represent x in formula (default: "X").
Only one character allowed. Only used when calculating SS from raw data
(\code{x}).}

\item{SS.f}{Formula to use for sum of squares calculation (either
solve_sum_squares or solve_sum_squares2). Only used when calculating SS
from raw data (\code{x}).}

\item{abbrev_sum}{Numeric scalar. Maximum length of x before it abbreviates
explicit summation within the solution using an ellipsis? (See
\code{\link{summation}}.) Only used when calculating SS from raw data
(\code{x}).}

\item{...}{Additional arguments to override default behaviors (see
\code{\link{handcalcs_defaults}}).}
}
\value{
\code{solve_sd()} returns a list with the interim values and calculations
(\code{x}, \code{SS}, \code{n}, \code{s2}), the final value (\code{s}), the solution string
(\code{solution}), and the bare formula (\code{formula}) in LaTeX format.
\code{sigma2_formula} returns just the bare formula in LaTeX format as a
character string.
}
\description{
Calculates sample standard deviation (\eqn{s = \sqrt[SS / (n - 1)]}) either from a set of
raw values (\code{x}) OR from the sum of squares (\code{SS}) and sample size
(\code{n}), OR from the sample variance (\code{s2}). If providing raw
data, it will include the calculation of SS in the solution string. If
providing \code{SS} and \code{n}, it will include the calculation of the
sample variance in the solution string. Otherwise, if only \code{s2}
is provided, will simply show the final calculation (\eqn{\sqrt(s^2)}).
}
\examples{
# Simple usage: generate raw data, produce results
(x <- sample(x = 1:10, size = 20, replace = TRUE))
solve_s(x = x, sub = 'X')

# Alternatively, provide SS and n
solve_s(SS = 100, n = 10)

# Or, simply provide s2
solve_s(s2 = 10)

# Must provide enough values for calculation:
\dontrun{
solve_s(SS = 100)
}

# If you just want the bare formula as a string, use a formula function():
s_formula()

# Can set parameters to change symbols used:
s_formula(sub = "Y")
}
