% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_values.R
\name{solve_t_to_p}
\alias{solve_t_to_p}
\title{Calculate \emph{p} values from the \emph{t} distribution.}
\usage{
solve_t_to_p(t, df, direction = c("pos", "neg", "both"), ...)
}
\arguments{
\item{t}{Numeric scalar. Value of t.}

\item{df}{Numeric scalar. Degrees of freedom.}

\item{direction}{For which tail(s) of the distribution should the p-value be
located: pos (area under the curve above \code{t}), neg (area under the curve
below \code{t}), both (area under the curve above +\code{t} and below -\code{t}).}

\item{...}{Additional arguments to override default behaviors (see
\code{\link[=handcalcs_defaults]{handcalcs_defaults()}}}
}
\value{
Returns a list with the provided values (\code{t}, \code{df}, \code{direction}), the
final value (\code{p}), and the solution string (\code{solution}) in LaTeX format.
}
\description{
Calculate \emph{p} values from the \emph{t} distribution.
}
\examples{
solve_t_to_p(1.96, direction = 'both')

# In some cases, you may need to change the default value of round_final
solve_t_to_p(1.645, direction = 'pos', round_final = 3)

}
